+++ example/lib/main.dart
@@ -15,7 +15,8 @@ import 'package:plex/plex_widget.dart';
 import 'package:plex/plex_widgets/plex_adv_data_table.dart';
 import 'package:plex/plex_widgets/plex_data_table.dart';
 import 'package:plex/plex_widgets/plex_data_table_paginated.dart';
-import 'package:plex/plex_widgets/plex_input_widget_2.dart';
+import 'package:plex/plex_widgets/plex_form.dart';
+import 'package:plex/plex_widgets/plex_form_field_widgets.dart';
 import 'package:plex_app/screens/home_screen.dart';
 import 'package:plex_app/screens/second_screen.dart';
 
@@ -432,14 +433,11 @@ void main() async {
           selectedLogo: const Icon(Icons.panorama_photosphere_select),
           screen: (context, {data}) => Column(
             children: [
-              PlexInputWidget2.input(
-                title: "Input",
-                helperText: "Input Usage",
-              ),
-              PlexInputWidget2.dropdown(
-                title: "Input",
-                helperText: "Input Usage",
-                dropdownItems: const [1,12,13,14,23,5,3,12,55,65],
+              PlexFormFieldInput(
+                properties: const PlexFormFieldGeneric.title("Field One"),
+                inputOnChange: (value) {
+
+                },
               ),
             ],
           ),